# perf: improve TypeScript's typecheck performance