--- a/specifyweb/frontend/js_src/lib/components/DataModel/businessRuleDefs.ts
+++ b/specifyweb/frontend/js_src/lib/components/DataModel/businessRuleDefs.ts
@@ -1,4 +1,4 @@
-import type { RA } from '../../utils/types';
+import type { RA, RR } from '../../utils/types';
 import type { BusinessRuleResult } from './businessRules';
 import type { AnySchema, TableFields } from './helperTypes';
 import {
@@ -65,6 +65,10 @@ export type UniquenessRule<SCHEMA extends AnySchema> = {
 type MappedBusinessRuleDefs = {
   readonly [TABLE in keyof Tables]?: BusinessRuleDefs<Tables[TABLE]>;
 };
+type GenericBusinessRuleDefs = RR<
+  keyof Tables,
+  BusinessRuleDefs<AnySchema> | undefined
+>;
 
 export const nonUniqueBusinessRuleDefs: MappedBusinessRuleDefs = {
   BorrowMaterial: {
@@ -357,42 +361,40 @@ export const nonUniqueBusinessRuleDefs: MappedBusinessRuleDefs = {
   },
 };
 
-/*
- * From this code, Typescript believes that a businessRuleDefs uniqueIn can be
- * from any table
- *
- * For example, it believes the following is possible:
- * BusinessRuleDefs<BorrowMaterial> & {uniqueIn: UniquenessRule<Accession> | UniquenessRule<AccessionAgent> | ...}
- */
-// @ts-expect-error
-export const businessRuleDefs: MappedBusinessRuleDefs = Object.fromEntries(
-  (
-    Object.keys({ ...uniqueRules, ...nonUniqueBusinessRuleDefs }) as RA<
-      keyof Tables
-    >
-  ).map((table) => {
-    /*
-     * To ensure compatibility and consistency with other areas of the frontend,
-     * the undefined type is preferable over the null type.
-     * In the JSON uniqueness rules, if a field should be unique at a global (institution)
-     * level, then it is unique in 'null'.
-     * Thus we need to replace null with undefined
-     */
-    const uniquenessRules: UniquenessRule<Tables[typeof table]> | undefined =
-      uniqueRules[table] === undefined
-        ? undefined
-        : Object.fromEntries(
-            Object.entries(uniqueRules[table]!).map(([fieldName, rule]) => [
-              fieldName,
-              rule[0] === null ? [undefined] : rule,
-            ])
-          );
-    const ruleDefs =
-      nonUniqueBusinessRuleDefs[table] === undefined
-        ? uniquenessRules === undefined
+const genericNonUniqueRuleReferences =
+  nonUniqueBusinessRuleDefs as GenericBusinessRuleDefs;
+
+const genericBusinessRuleDefs: GenericBusinessRuleDefs = Object.fromEntries(
+  Object.keys({ ...uniqueRules, ...genericNonUniqueRuleReferences }).map(
+    (table) => {
+      /*
+       * To ensure compatibility and consistency with other areas of the frontend,
+       * the undefined type is preferable over the null type.
+       * In the JSON uniqueness rules, if a field should be unique at a global (institution)
+       * level, then it is unique in 'null'.
+       * Thus we need to replace null with undefined
+       */
+      const uniquenessRules: UniquenessRule<Tables[typeof table]> | undefined =
+        uniqueRules[table] === undefined
           ? undefined
-          : { uniqueIn: uniquenessRules }
-        : { ...nonUniqueBusinessRuleDefs[table], uniqueIn: uniquenessRules };
-    return [table, ruleDefs];
-  })
+          : Object.fromEntries(
+              Object.entries(uniqueRules[table]!).map(([fieldName, rule]) => [
+                fieldName,
+                rule[0] === null ? [undefined] : rule,
+              ])
+            );
+      const ruleDefs: BusinessRuleDefs<AnySchema> | undefined =
+        genericNonUniqueRuleReferences[table] === undefined
+          ? uniquenessRules === undefined
+            ? undefined
+            : { uniqueIn: uniquenessRules }
+          : {
+              ...genericNonUniqueRuleReferences[table],
+              uniqueIn: uniquenessRules,
+            };
+      return [table, ruleDefs];
+    }
+  )
 );
+export const businessRuleDefs =
+  genericBusinessRuleDefs as MappedBusinessRuleDefs;