--- a/specifyweb/frontend/js_src/lib/components/PickLists/definitions.ts
+++ b/specifyweb/frontend/js_src/lib/components/PickLists/definitions.ts
@@ -14,7 +14,7 @@ import { getField } from '../DataModel/helpers';
 import type { SerializedResource, TableFields } from '../DataModel/helperTypes';
 import type { SpecifyResource } from '../DataModel/legacyTypes';
 import { deserializeResource } from '../DataModel/serializers';
-import { tables } from '../DataModel/tables';
+import { genericTables, tables } from '../DataModel/tables';
 import type { PickList, PickListItem, Tables } from '../DataModel/types';
 import { hasToolPermission } from '../Permissions/helpers';
 
@@ -100,7 +100,7 @@ export function definePicklist(
 export const pickListTablesPickList = f.store(() =>
   definePicklist(
     '_TablesByName',
-    Object.values(tables).map(({ name, label }) =>
+    Object.values(genericTables).map(({ name, label }) =>
       createPickListItem(name.toLowerCase(), label)
     )
   )
@@ -147,7 +147,7 @@ export const getFrontEndPickLists = f.store<{
   // Like pickListTablesPickList, but indexed by tableId
   const tablesPickList = definePicklist(
     '_Tables',
-    Object.values(tables).map(({ tableId, label }) =>
+    Object.values(genericTables).map(({ tableId, label }) =>
       createPickListItem(tableId.toString(), label)
     )
   );