In https://github.com/microsoft/TypeScript/wiki/Performance#extendeddiagnostics, I saw that you can run this command to see what step of TypeScript takes the most time:

```sh
npx tsc --extendedDiagnostics
```

That told me that out of 17s it takes to run TypeScript on our codebase (on m1 mac), 16s was spent checking types.

I wanted to see if there places where we could optimize code a bit to simplify the job for TypeScript

Following instructions in https://github.com/microsoft/TypeScript/wiki/Performance-Tracing I run this:

```sh
npx tsc --generateTrace ~/Downloads/nameOfDirectoryToCreate --incremental false
```

That generated a performance trace report.

Then, I uploaded that trace (~/Downloads/nameOfDirectoryToCreate/trace.json) into https://www.speedscope.app/ and enabled Left heavy view. 

![Screenshot 2023-11-29 at 21 48 10](https://github.com/specify/specify7/assets/40512816/18a87b6b-8396-4469-91aa-913b51144fd0)


That showed that nearly 4s is spent in a single place:

https://github.com/specify/specify7/blob/602fd7d9bd2fd67d4a2fa0aa3b523f789775faca/specifyweb/frontend/js_src/lib/components/SchemaConfig/Tables.tsx#L87-L96

The issue is that by calling `Object.values(tables)`, we are getting a giant union of very different types (each table has a very different type because the type is `RA<SpecifyTables<keyof Tables>>`). So then when we try to .filter() over it, TypeScript has to do a lot of work making sure each union member has some common subtype.

The solution is to simplify the above type to `RA<SpecifyTables<AnySchema>>` - in this case, we don't care about each individual table schema, but only care about the fact that tables have "isSystem" and "name" fields.

---

Then I noticed many more places in the code where we do `Object.values(tables)`, `Object.entries(tables)` or `tables[tableName]` where `tableName` is `keyof Tables` - in all these cases, a giant union is created, but we don't care about a type of any particular table, but about the aggregate type common to all tables. Thus I introduced a `genericTables` table, which has type `RR<keyof Tables, SpecifyTable<AnySchema>>`.

tables should be used when you care about a particular table `table.Accession` or set of tables `table['${treeName}TreeDef']`.
If you don't care about the specifics of each individual table, but care about tables in general, use `genericTables` instead.

---

Then I went over few more places in the code and similarly resolved identified issues (had similar problem in business rule defs)

---

At the end, type check time went down from 16s to 9s.
While the biggest bottlenecks have been addressed, if someone wants to spend a bit more time on this, we could get it down by ~4s more as there are still many candidates for optimization:

![Screenshot 2023-11-29 at 21 56 45](https://github.com/specify/specify7/assets/40512816/40672621-10b6-4902-8ddf-2394b23307e4)